requirements.txt
chardet==2.3.0
colorama==0.3.2
defer==1.0.6
Django==1.9.8
djangorestframework==3.4.1
html5lib==0.999
ndg-httpsclient==0.3.2
psutil==2.1.1
pyasn1==0.1.7
pygobject==3.14.0
pyOpenSSL==0.13.1
pysqlite==2.6.3
python-apt===0.9.3.11build1
python-debian==0.1.22
pyxdg==0.25
requests==2.4.3
six==1.9.0
urllib3==1.9.1
virtualenv==15.0.2
